# swagger/swagger.json
{"swagger":"2.0","info":{"title":"Data client to add or delete data","description":"Add or delete data","version":""},"host":"localhost:2626","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/data/add/{hash}/{private_key}":{"post":{"tags":["DataClient"],"summary":"add DataClient","description":"add data hash","operationId":"DataClient#add","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"hash","in":"path","description":"data IPFS address","required":true,"type":"string"},{"name":"private_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/data/agree/{ETH_key}/{request_id}":{"post":{"tags":["DataClient"],"summary":"agree DataClient","description":"agree data request for request [ID]","operationId":"DataClient#agree","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"ETH_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"},{"name":"request_id","in":"path","description":"request[ID]","required":true,"type":"integer"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/data/askComputing/{hash}/{ETH_key}/{request_id}":{"post":{"tags":["DataClient"],"summary":"askComputing DataClient","description":"ask for computing for [request_id] on computing resourse[hash]","operationId":"DataClient#askComputing","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"ETH_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"},{"name":"hash","in":"path","description":"computing resourse hash","required":true,"type":"string"},{"name":"request_id","in":"path","description":"request[ID]","required":true,"type":"integer"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/data/del/{hash}/{private_key}":{"post":{"tags":["DataClient"],"summary":"del DataClient","description":"delete data hash","operationId":"DataClient#del","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"hash","in":"path","description":"data IPFS address","required":true,"type":"string"},{"name":"private_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/data/upload/{hash}/{ETH_key}/{request_id}":{"post":{"tags":["DataClient"],"summary":"uploadData DataClient","description":"upload encrypted data[hash] for [request_id]","operationId":"DataClient#uploadData","produces":["application/vnd.goa.error","plain/text"],"parameters":[{"name":"ETH_key","in":"path","description":"ETH private key for transaction","required":true,"type":"string"},{"name":"hash","in":"path","description":"encrypted data hash","required":true,"type":"string"},{"name":"request_id","in":"path","description":"request[ID]","required":true,"type":"integer"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/swagger-ui/{filepath}":{"get":{"summary":"Download swagger-ui/","operationId":"swagger-ui#/swagger-ui/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/swagger.json":{"get":{"summary":"Download swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}}},"definitions":{"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}}}